Refactor and style code. Add migrations files
