+++ estela-api/api/serializers/job.py
@@ -1,5 +1,5 @@
-from rest_framework import serializers
 from django.shortcuts import get_object_or_404
+from rest_framework import serializers
 
 from api import errors
 from api.serializers.job_specific import (
@@ -7,7 +7,7 @@ from api.serializers.job_specific import (
     SpiderJobEnvVarSerializer,
     SpiderJobTagSerializer,
 )
-from api.utils import update_stats_from_redis, delete_stats_from_redis
+from api.utils import delete_stats_from_redis, update_stats_from_redis
 from config.job_manager import job_manager
 from core.models import (
     DataStatus,
