+++ estela-api/api/serializers/spider.py
@@ -1,9 +1,9 @@
 from rest_framework import serializers
 
-from api.serializers.job_specific import SpiderJobEnvVarSerializer
-from api.views.project import update_env_vars
-from core.models import DataStatus, Spider
 from api import errors
+from api.serializers.job_specific import SpiderJobEnvVarSerializer
+from api.utils import update_env_vars
+from core.models import DataStatus, Spider
 
 
 class SpiderSerializer(serializers.ModelSerializer):
@@ -17,12 +17,19 @@ class SpiderSerializer(serializers.ModelSerializer):
     )
 
     env_vars = SpiderJobEnvVarSerializer(
-        many=True, required=False, help_text="Spider env variables."
+        many=True, required=False, help_text="Spider environment variables."
     )
 
     class Meta:
         model = Spider
-        fields = ("sid", "name", "project", "env_vars", "data_status", "data_expiry_days")
+        fields = (
+            "sid",
+            "name",
+            "project",
+            "env_vars",
+            "data_status",
+            "data_expiry_days",
+        )
 
 
 class SpiderUpdateSerializer(serializers.ModelSerializer):
@@ -30,7 +37,7 @@ class SpiderUpdateSerializer(serializers.ModelSerializer):
         read_only=True, help_text="A UUID identifying this spider."
     )
     env_vars = SpiderJobEnvVarSerializer(
-        many=True, required=False, help_text="Project env variables."
+        many=True, required=False, help_text="Project environment variables."
     )
     data_status = serializers.ChoiceField(
         choices=DataStatus.HIGH_LEVEL_OPTIONS,
@@ -47,10 +54,10 @@ class SpiderUpdateSerializer(serializers.ModelSerializer):
         fields = ("sid", "env_vars", "data_status", "data_expiry_days")
 
     def update(self, instance, validated_data):
-        data_status = validated_data.get("data_status", "")
+        data_status = validated_data.get("data_status", None)
         data_expiry_days = validated_data.get("data_expiry_days", 1)
         env_vars = validated_data.get("env_vars", [])
-        if "data_status" in validated_data:
+        if data_status:
             instance.data_status = data_status
             if data_status == DataStatus.PENDING_STATUS and data_expiry_days > 0:
                 instance.data_expiry_days = data_expiry_days
